==> notebooks/1 - MongoQueries_CompaniesCollection_find_city.ipynb <==
## 1 - Mongo DB queries from "companies" dataset.
#### Find succesfull companies in the GAMING related industry (requirements marked in bold above)

- NEED 1: The CEO is vegan = 1 --> vegan restaurants

- NEED 2: The office dog—"Dobby" needs a hairdresser every month. Ensure there's one not too far away = 87 --> vet or pet hairdresser

- NEED 3: Everyone in the company is between 25 and 40, give them some place to go party = 87  --> pubs and clubs

- NEED 4: 30% of the company staff have at least 1 child = 26 --> nurseries or schools

#### - NEED 5: Designers like to go to design talks and share knowledge. There must be some nearby companies that also do design. = 20 --> gaming companies around

- NEED 6: Account managers need to travel a lot = 20 --> airports or train stations

#### - NEED 7: Developers like to be near successful tech startups that have raised at least 1 Million dollars = 15 (Front and Back) --> succesfull tech companies around

### Import libraries

from pymongo import MongoClient
import pandas as pd
import numpy as np
import json
from pandas import json_normalize
import matplotlib.pyplot as plt

import sys
sys.path.append('../')
import src.offices_functions as of

import warnings
warnings.filterwarnings('ignore')



# https://www.gameindustrycareerguide.com/best-cities-for-video-game-development-jobs/

### MongoCompass connection

client = MongoClient("mongodb://localhost/companies")
db = client.get_database()

list(db.list_collections())

db.list_collection_names()

collection = db.get_collection("companies")
collection

c = collection

### Format offices with several locations

res = c.aggregate([{"$unwind":"$offices"},{"$project":{"_id":0}}])

offices = client.companies.offices
offices.insert_many(res)

res = c.aggregate([
    {"$unwind":"$offices"},
    {"$match":{"offices.latitude":{"$ne":None}, "offices.longitude":{"$ne":None}}}, # # Give only companies where "offices.latitude" and "offices.longitude" is not None
    {"$project":{"_id":0}}])

offices.drop()
offices.insert_many(res)

res = list(offices.find({},{"offices":1}))

### Companies Search

#### Since the company is in the GAMING industry, employees might be wanted to be close to this other type of companies: games_video and companies in tech

c.distinct("category_code")

res = list(offices.find({},{"offices":1}))

# select game_video companies
category_1 = {"category_code": "games_video"}
 

# select companies ended in "tech" companies
category_2 = {"category_code":{"$regex": ".*tech.*"}}

              
# give me only companies that total_money_raised is greater than 1M(€/$)"
cond_1 = {"total_money_raised":{"$regex": "M$"}}
              
# give me only companies with "offices.city"
cond_2 = {"offices.city": {"$ne":None}}
cond_3 = {"offices.city": {"$ne":""}}


# we can to make visival the following attributes:
categ_proj = {"_id":1, "name":1, "offices":1, "category_code":1}

affinity_companies = list(offices.find({"$and":[{"$or":[category_1, category_2]}, cond_1, cond_2, cond_3]}, categ_proj).sort("total_money_raised",-1))

affinity_companies[:5]

# Convert to a DataFrame
df_affinity_companies = pd.DataFrame(affinity_companies)

df_affinity_companies.head()

# Get the info within office dicts into column

column_dict_keys = ["address1", "address2", "city", "latitude", "longitude", "country_code"]

of.getFromDictColumn("offices", column_dict_keys, df_affinity_companies)

# Create a subset of companies count by country
df_affinity_companies_top_countries =  df_affinity_companies.groupby(["country_code"] ).size().to_frame(name = "companies_count").reset_index()
df_top3_countries = df_affinity_companies_top_countries.nlargest(3,"companies_count")
df_top3_countries

# Create a subset of companies count by city

df_affinity_companies_top_cities =  df_affinity_companies.groupby(["city", "country_code" ] ).size().to_frame(name = "companies_count").reset_index()
df_top3_cities = df_affinity_companies_top_cities.nlargest(3,"companies_count")
df_top3_cities

#### This resut tell us that the gaming and tech brain is in USA, precisely in San Francisco and New York 

### Select randomly a company address from each city

#### Address in San Franciso

df_affinity_companies_subset = df_affinity_companies[["name", "address1", "city", "latitude", "longitude", "country_code"]]

df_san_francisco_companies = df_affinity_companies_subset[(df_affinity_companies_subset["city"] == "San Francisco") & (df_affinity_companies_subset["country_code"] == "USA")]

df_san_francisco_company = df_san_francisco_companies.sample()
pd.set_option("max_colwidth", 80)
df_san_francisco_company

#### Address in New York

df_new_york_companies = df_affinity_companies_subset[(df_affinity_companies_subset["city"] == "New York") & (df_affinity_companies_subset["country_code"] == "USA")]

df_new_york_company = df_new_york_companies.sample()
df_new_york_company

df_affinity_companies.to_csv("affinity_companies.csv", index=0)

#### These three cities will meet two of our employees requirementes:
#### - NEED 5: Designers like to go to design talks and share knowledge. There must be some nearby companies that also do design. = 20 --> gaming companies around
#### - NEED 7: Developers like to be near successful tech startups that have raised at least 1 Million dollars = 15 (Front and Back) --> succesfull tech companies around
