==> notebooks/2 - Foursquare API.ipynb <==
# 2 - APIs Geocode and Foursquare

### Import Libraries

import requests 
import json
from pandas import json_normalize
import pyjsonviewer
from dotenv import load_dotenv
import os
import pandas as pd
from functools import reduce
import operator
import geopandas as gdp
import shapely

from pymongo import MongoClient

import sys
sys.path.append('../')
import src.api_functions as af
import src.offices_functions as of


import warnings
warnings.filterwarnings('ignore')

- NEED 1: The CEO is vegan = 1 --> vegan restaurants

- NEED 2: The office dog—"Dobby" needs a hairdresser every month. Ensure there's one not too far away = 87 --> vet or pet hairdresser

- NEED 3: Everyone in the company is between 25 and 40, give them some place to go party = 87  --> pubs and clubs

- NEED 4: 30% of the company staff have at least 1 child = 26 --> nurseries or schools

- NEED 5: Designers like to go to design talks and share knowledge. There must be some nearby companies that also do design. = 20 --> gaming companies around

- NEED 6: Account managers need to travel a lot = 20 --> airports or train stations

- NEED 7: Developers like to be near successful tech startups that have raised at least 1 Million dollars = 15 (Front and Back) --> succesfull tech companies around

load_dotenv()

client_id = os.getenv("token1")
client_secret = os.getenv("token2")

### We get the coordenates the selected cities

df_affinity_companies = pd.read_csv("affinity_companies.csv")
df_affinity_companies_top_cities =  df_affinity_companies.groupby(["city", "country_code" ] ).size().to_frame(name = "companies_count").reset_index()
df_top2_cities = df_affinity_companies_top_cities.nlargest(2,"companies_count")
df_top2_cities

### San Francisco address

df_affinity_companies_subset = df_affinity_companies[["name", "address1", "city", "latitude", "longitude", "country_code"]]
df_san_francisco_address = df_affinity_companies_subset[(df_affinity_companies_subset["city"] == "San Francisco") & (df_affinity_companies_subset["country_code"] == "USA") & (df_affinity_companies_subset["name"] == "PlayFirst")]

df_san_francisco_address

### New York address

df_affinity_companies_subset = df_affinity_companies[["name", "address1", "city", "latitude", "longitude", "country_code"]]
df_new_york_address = df_affinity_companies_subset[(df_affinity_companies_subset["city"] == "New York") & (df_affinity_companies_subset["country_code"] == "USA") & (df_affinity_companies_subset["name"] == "Major League Gaming")]

df_new_york_address

top_cities_address = ["160 Spear St San Francisco USA", "3 Park Avenue New York USA"]

coords = af.getCoordinates(top_cities_address)

for k, v in coords.items():
    print(v.keys())

print(coords.items())

coordinates = af.actualCoordinates(coords)

coordinates

### FOURSQUARE API CALLING

We are trying to meet now requirements in bold:

#### - NEED 1: The CEO is vegan = 1 --> vegan restaurants

#### - NEED 2: The office dog—"Dobby" needs a hairdresser every month. Ensure there's one not too far away = 87 --> vet or pet hairdresser

#### - NEED 3: Everyone in the company is between 25 and 40, give them some place to go party = 87  --> pubs and clubs

#### - NEED 4: 30% of the company staff have at least 1 child = 26 --> nurseries or schools

- NEED 5: Designers like to go to design talks and share knowledge. There must be some nearby companies that also do design. = 20 --> gaming companies around

#### - NEED 6: Account managers need to travel a lot = 20 --> airports or train stations

- NEED 7: Developers like to be near successful tech startups that have raised at least 1 Million dollars = 15 (Front and Back) --> succesfull tech companies around

queries = ["Nursery", "Preschool", "Train Station", "Airport", "Pub", "Club", "Vet", "Pet Grooming", "vegan restaurant"]

url = "https://api.foursquare.com/v2/venues/explore"

San_Francisco_USA = af.foursquareQueries(coordinates["160 Spear St San Francisco USA"][1],coordinates["160 Spear St San Francisco USA"][0], url, *queries)

# San_Francisco_USA

New_York_USA = af.foursquareQueries(coordinates["3 Park Avenue New York USA"][1],coordinates["3 Park Avenue New York USA"][0], url, *queries)

###  Foursquare to Mongo

# with open('data.json', 'w') as f:
#     json.dump(resp, f)

# pyjsonviewer.view_data(json_data=resp)

San_Francisco_USA2 = af.responseGroups(San_Francisco_USA)

New_York_USA2 = af.responseGroups(New_York_USA)

# New_York_USA2

# def getFromDict(diccionario, mapa):
#     return reduce(operator.getitem,mapa,diccionario)

# def type_point(lista):
#     return {"type": "Point", 
#             "coordinates": lista }

### Dictionaries to dataframe

df_san_francisco_usa = af.toDataframe(*San_Francisco_USA2) 

df_san_francisco_usa.head(1)

# Get key name from category

df_san_francisco_usa_final = json_normalize(df_san_francisco_usa["category"])

df_san_francisco_usa_final.head(2)

df_san_francisco_usa_final.rename(columns = {0:'category_all'}, inplace = True)

# df_san_francisco_usa_final.head()

column_dict_keys = ["name"]

of.getFromDictColumn("category_all", column_dict_keys, df_san_francisco_usa_final)

df_san_francisco_usa_final2 = df_san_francisco_usa_final.drop(columns=['category_all'])
df_san_francisco_usa2 = df_san_francisco_usa.drop(columns=['category'])

df_san_francisco_usa_final2.rename(columns = {"name":"category"}, inplace = True)
df_san_francisco_usa_final2.head(2)

df_san_francisco_usa_final3 = pd.concat([df_san_francisco_usa2, df_san_francisco_usa_final2], axis=1)

df_san_francisco_usa_final3.head(2)

df_new_york_usa = af.toDataframe(*New_York_USA2)

df_new_york_usa.head(1)

# Get key name from category

df_new_york_usa_final = json_normalize(df_new_york_usa["category"])

df_new_york_usa_final.head(2)

df_new_york_usa_final.rename(columns = {0:'category_all'}, inplace = True)

of.getFromDictColumn("category_all", column_dict_keys, df_new_york_usa_final)

df_new_york_usa_final2 = df_new_york_usa_final.drop(columns=['category_all'])
df_new_york_usa2 = df_new_york_usa.drop(columns=['category'])

df_new_york_usa_final2.rename(columns = {"name":"category"}, inplace = True)
# df_new_york_usa_final2.head(2)

df_new_york_usa_final3 = pd.concat([df_new_york_usa2, df_new_york_usa_final2], axis=1)
df_new_york_usa_final3.head(2)

### Dataframes to csv

df_san_francisco_usa_final3.to_csv("../data/API_queries/san_francisco_usa.csv", index = False)

df_new_york_usa_final3.to_csv("../data/API_queries/new_york_usa.csv", index = False)
