==> notebooks/3 - Near Geoqueries.ipynb <==
# 3 - Location to Geopoints and Near Geoqueries

## Libraries

import pandas as pd
from bson.json_util import dumps
import shapely
import geopandas as gdp
# import json
from pymongo import MongoClient
from pymongo import GEOSPHERE

import sys
sys.path.append('../')
import src.pointype_queries_functions as ptf


import warnings
warnings.filterwarnings('ignore')

# Import API queries from each city

san_francisco_usa = pd.read_csv("../data/API_queries/san_francisco_usa.csv")
new_york_usa = pd.read_csv("../data/API_queries/new_york_usa.csv")

### Change coodinates to point type

gdf_san_francisco_usa = ptf.typePoint(san_francisco_usa)

gdf_san_francisco_usa.head(2)

gdf_new_york_usa = ptf.typePoint(new_york_usa)

gdf_new_york_usa.head(2)

### Gets queries, distances from venues to our location and creates city collections in our Mongo Database

#### San Francisco: 160 Spear St

lon, lat = -122.39312, 37.79179
max_distance = 6000
collection = "san_francisco"
mongo_database = "cities_research"

gdf_san_francisco_usa2 = ptf.nearQueries(gdf_san_francisco_usa, mongo_database, collection, lon, lat, max_distance)

# count results from query
ptf.countQueryDocs("san_francisco","cities_research")

# Creates dataframe
df_san_francisco_usa = pd.DataFrame(gdf_san_francisco_usa2)
df_san_francisco_usa.head(2)

#### New York: 3 Park Avenue

lon, lat = -73.98037, 40.69651
max_distance = 6000
collection = "new_york"
mongo_database = "cities_research"

gdf_new_york_usa2 = ptf.nearQueries(gdf_new_york_usa, mongo_database, collection, lon, lat, max_distance)

ptf.countQueryDocs(collection, mongo_database)

# Creates dataframe

df_new_york_usa = pd.DataFrame(gdf_new_york_usa2)
df_new_york_usa.head(2)

### Export to csv

df_san_francisco_usa.to_csv("../data/Near_geoqueries/near_san_francisco.csv", index = False)

df_new_york_usa.to_csv("../data/Near_geoqueries/near_new_york.csv", index = False)
