==> notebooks/4 - Compare cities.ipynb <==
## Libraries

import pandas as pd
import geopandas as gdp
import folium
from folium import Choropleth, Circle, Marker, Icon, Map
from folium.plugins import HeatMap, MarkerCluster
import re

import sys
sys.path.append('../')
import src.compare_cities_functions as ccf


import warnings
warnings.filterwarnings('ignore')

df_san_francisco_usa = pd.read_csv("../data/Near_geoqueries/near_san_francisco.csv")

df_new_york_usa = pd.read_csv("../data/Near_geoqueries/near_new_york.csv")

## Checking categories Foursquare found

df_san_francisco_usa.category.unique()

df_new_york_usa.category.unique()

## Categories consolidation

#### We want to consolidate the categories found into our main categories

- NEED 1: The CEO is vegan = 1 --> Vegan Restaurants

- NEED 2: The office dog—"Dobby" needs a hairdresser every month. Ensure there's one not too far away = 87 --> Pet Services (vet or pet hairdresser)

- NEED 3: Everyone in the company is between 25 and 40, give them some place to go party = 87  --> Clubbing (pubs, bars and clubs)

- NEED 4: 30% of the company staff have at least 1 child = 26 --> Preschool (nurseries or schools)


- NEED 6: Account managers need to travel a lot = 20 --> Transportation (airports or train stations)

df_san_francisco_usa["category_simplify"] = df_san_francisco_usa["category"]
df_new_york_usa["category_simplify"] = df_new_york_usa["category"]

ccf.categoryAggr(df_san_francisco_usa, "category_simplify")

ccf.categoryAggr(df_new_york_usa, "category_simplify")

#### We've got a few categories results that are not of our interest. We delete those results

main_requirements_list = ['Preschool', 'Going Out', 'Vegan Restaurant', 'Transportation', 'Pet Services']

# For San Francisco
df_san_francisco_usa.drop(df_san_francisco_usa.loc[~df_san_francisco_usa['category_simplify'].isin(main_requirements_list)].index, inplace=True)

df_san_francisco_usa.category_simplify.unique()

# For New York
df_new_york_usa.drop(df_new_york_usa.loc[~df_new_york_usa['category_simplify'].isin(main_requirements_list)].index, inplace=True)

df_new_york_usa.category_simplify.unique()

#### We can now put the two dataframes together

df_san_francisco_usa["city"] = "San Francisco"
df_new_york_usa["city"] = "New York"

df_san_francisco_usa.shape

df_new_york_usa.shape

df_cities = pd.concat([df_san_francisco_usa, df_new_york_usa], axis=0)

df_cities.shape

df_cities.head(2)

## Calculate City Ranking

#### Normalize "distance"

Data Normalization is a common practice which consists of transforming numeric columns to a common scale. For this case we are going to use the min-max scaling

df_cities_mean = df_cities.groupby(["category_simplify", "city"]).mean()

df_cities_mean.reset_index(drop = False)

df_cities_normalized = ccf.min_max_scaling(df_cities_mean, "distance")

df_cities_normalized["importance"] = df_cities_normalized["category_simplify"]

ccf.importance(df_cities_normalized, "importance")

df_cities_normalized.head()

# df_cities_normalized.info()

### Weights

- Vegan Restaurant = 0.25
- Pet Services = 0.25
- Going Out = 0.2
- Preschool = 0.15
- Transportation = 0.15 

df_cities_normalized["ranking"] = df_cities_normalized["distance"] * df_cities_normalized["importance"]

df_cities_normalized2 = df_cities_normalized.groupby(["city"]).sum()

df_cities_normalized2.sort_values("ranking")

### We have a winner: [Fake Tales Of San Francisco](https://www.youtube.com/watch?v=ePg1tbia9Bg)

# export dataframe
df_san_francisco_usa.to_csv("../data/San_Francisco_map_ready/san_francisco_final.csv", index = False)
